Refactor subscriptionService
