+++ lib/app/modules/root/root_binding.dart
@@ -1,5 +1,4 @@
 import 'package:get/get.dart';
-import 'package:habitat_ft_user/app/modules/home/services/subscription_service.dart';
 import 'package:habitat_ft_user/app/modules/login/services/auth_service.dart';
 
 import 'package:habitat_ft_user/app/modules/root/root_controller.dart';
@@ -9,6 +8,5 @@ class RootBinding extends Bindings {
   void dependencies() {
     Get.put<RootController>(RootController());
     Get.put<AuthService>(AuthService());
-    Get.put<SubscriptionService>(SubscriptionService());
   }
 }
